refactor: extract the /token grant handlers into routes/oauth_grants.py (2.9 arch review)
oauth.py was 2296 lines; the token area (the five _grant_* handlers plus
_GrantOutcome/_GrantContext/_resolve_token_resource/GrantResult/_GRANT_HANDLERS)
is ~750 of them and a natural, cohesive unit. Move it verbatim to
routes/oauth_grants.py; oauth.py's token() imports _GRANT_HANDLERS,
_GrantContext and _GrantOutcome and dispatches exactly as before. oauth.py is
now 1550 lines.

Chosen over a full package split (arch review finding 3): the handlers depend
only on services/models, never back on routes.oauth, so the import is one-way
with no cycle and no new indirection - a small, mechanical move rather than a
2300-line reshuffle. The module's external contract is unchanged (oauth_bp),
and _GrantOutcome stays importable from routes.oauth for the one test that uses
it.

Behaviour-preserving. Verified: unit suite 1812, lint-imports contracts kept
(layering intact with the new module), ruff and mypy clean, and the default
e2e suite 60/60 - OAuth2/OIDC 26/26 exercises every moved grant handler
(authorization_code, client_credentials, refresh, password, PKCE, resources).
